+++ addOns/authhelper/CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ## Unreleased
 ### Fixed
 - Potential timing issue trying to use browser based auth to authenticate before the session management method has been identified.
+- Timing issue with session management detection.
 
 ## [0.13.0] - 2024-05-07
 ### Changed